Replace the fragile sleep() calls in AuthListenerTest with more robust mechanism to ensure that the authentication listeners have run to completion on both caller and receiver side.
The mechanism is to make a second (insecure) method call after the
first method call, then check the assertions.  The method reply is
guaranteed not to be received until authentication has completed on
the receiver side so we can be sure that by the time the first call
completes then authentication is completed on both sides.