Fixed issue that prevented the junit test running in the Android emulator.
The liballjoyn_java.so was being placed in the armeabi-v7a folder for the junit tests.  The test would run as expected on anything that used ARM7  but would not run in the emulator that uses ARM5. By placing the liballjoyn_java.so file in the armeabi folder it will run on all the phones and the emulator.