Initial update of junit tests for sessions.
The primary objective of this initial update was to get the junit code to compile.  Many of the tests that previously passed now fail.  All instances of TODO will be removed from the code as the junit test continue to be updated.  Code was commented out and not removed so it could be referenced while updating the code.