MBUS-410 Added check for ER_BUS_INTERFACE_ACTIVATED to JNI code when calling AddMember or AddProperty.
The junit tests are acting as both the service and the client and the sercvice and the clinet side of the tests both try to create a new interface with the same members and properties.  The old code would check for the error ER_BUS_MEMBER_ALREADY_EXISTS but this error is not was is reported now that the new error ER_BUS_INTERFACE_ACTIVATED has been added to the code.