Add a check in the BusAttachment.java code to check the os.type
Before assigning the default value for org.alljoyn.bus.address check the os.type if running in windows use tcp:addr for everything else use unix:abstract.  This makes it so the code will run in windows without having to pass in an option to set a bus address that will work in windows.