Added a ProgressDialog to the android client samples
The project samples did not give any indication that they were searching for the service.  With this addition a ProgressDialog will come up on the screen till the client returns from the joinsession call.  This prevents the user from pressing buttons and expecting feedback from the service till the service has been joined.