have scons build and install the JavaSDKDoc samples into the dist folder.
When the Guide to AllJoyn Development Using the Java SDK document was written a lot of sample code was written that was intended to be used as reference samples.  However these samples were not being added to the SDK distribution because the eclipse projects were not being built to verify they would build.  These scons scripts will create a java manifest file for each sample and use that to build a jar file for each sample that will be placed in the java/jar folder.  The script will also copy the samples to the java/samples folder so it can be found by people that only download the SDK and not the code from source.