[ALLJOYN-735,736] keep junit tests from failing on android and some linux builds
BusAttachment objects were not being garbage collected after each test run because the keyStoreListeners registered to the objects were not being deleted this caused the BusAttachments destructor to never be called.  Since many file descriptors are opened for each BusAttachment instance we quickly ran out of avalible BusAttahments avalible from the system.

AuthListenerTest still creates some file descriptors that are not being freed but its not as common as it was.