[ALLJOYN 728] Removed code that resulted in BusAttachmentTest reporting an error tests intermittently reported a DBUS_RELEASE_NAME_REPLY_NON_EXISTENT error that resulted from calling BusAttachment.releaseName after the name had already be free.
This code was originally used to make sure the name was free before the next test was run.  Now that the BusAttachement.release() method has been added  to the test tear down we don't have to worry about the name being owned by an object that has not yet been cleaned up by Java's garbage collector.