Add a new WiFi Direct version of the Simple Service
This app is nearly identical to the normal Simple Service, but is also
able to automatically approve WiFi Direct group formation.  It
contains a WifiDirectAutoAccept class that encapsulates the group
approval code, but this class depends on APIs found in Android API 16.
This service operates with the normal Simple Client.

Change-Id: Ifd051ce8e992f2a252148dd8fbd4dd644e17afa4