ALLJOYN-1651 Removing samples that no longer show proper usage of AllJoyn
The following samples were originally created to help with Java binding
development: AddressBook, Game, props, and simple.  The samples were never
updated with sessions and now that we have moved to the built in discovery.
There is no way the services can find the client.  The choice was to add
sessions to the samples or remove the samples.  The decision was to remove
the samples since we have similar samples in the JavaSDKDoc folder.

Change-Id: I60ae7cf12ade25a2fcc127e2d5a20b966944c259