ALLJOYN-1701 Modify android samples so that they work for latest Eclipse Juno and ADT plugin
The current android samples have had problems on different versions/combinations of Eclipse
and ADT plug-ins. With this change we want to make sure that, going forward, all the android samples
work on at least Eclipse Juno. More specifically, the eclipse ADT bundle that users download from
android developer webpage. The ADT version right now is 21.1 and the Eclipse version is 4.2.1 in
the bundle.