ALLJOYN-1821 Shutdown Hook added to the JavaSDKDoc samples
The native BusAttachment is not being garbage-collected before the
Java VM exits that was causing segfaults when the user terminated the
samples using kill commands like ctrl+c.

The Shutdown Hook calls BusAttachment.release() when SIGINT signal is
received.  This properly closes down the BusAttachment without needing
to rely on Java's garbage-collector.

This is a work around till we are able to fix the underlying issue.
See ALLJOYN-1761 for more on the VM garbage-collector issue.

Change-Id: I948c218e3ec4359e9a0fa815c52664db2a2d5f2a