ALLJOYN-2109 re-enabled ProxyBusObject tests that were commented out
Every single test in the ProxyBusObjectTest suit was commented out while work
was being done to add sessions to the Java language bindings in early 2011.

The tests were not re-enabled when the sessions related work was completed.
The empty tests were left running giving the false impression that the
ProxyBusObject class was still being tested.

This is a full rework of the existing tests.  Some of the tests required very
little change to get them working again.  Some of the tests were no longer
valid.  Some additional tests were added to cover the classes that have been
added or did not already have a unit test.

Change-Id: I073290b3296fad2a4e19aa5a2179704b92103c54