ALLJOYN-???? Fixed failure in ObjectSecurity Tests on Android
On Android the new ObjectSecurity tests fail when they try to create the
KeyStore file. This is because the default location for creating a keystore
file is at the location specified by the $HOME environment variable or if
that is undefined at '/' root.  On android $HOME is not defined and it is
illegal to create any files at root.  Since we cannot create a file we get
the ER_BUS_WRITE_ERROR from the KeyStore.StoreRequest of the
DefaultKeyStoreListener.

To solve the problem I used the same solution employed by the AuthListenerTest
Create custom KeyStoreListener that stores the KeyStore in memory. This change
adds an in memory  KeyStoreListener and registers it in place of the default
KeyStore that was causing problems on Android.

Change-Id: Ie9cde66d22d1c31bb8a2ef8d951e1a7c8ea90ba5