ALLJOYN-2345 Encryption is applied when using properties on a secure interface
Code has been added to the methodCall code in the jni code that will try and
determine if encryption should be used when doing DBus Properties method call
it does this using the interface name sent in when making the properties method
call.  If the code is unable to find the interface to obtain its security it is
possible for a properties method call to still fail.

New methods have been added to the ProxyBusObject class that mirrors the same
methods found in the C++ code.
   - ProxyBusObject.getProperty
   - ProxyBusObject.setProperty
   - ProxyBusObject.getAllProperties

Since these calls are made from the same interface that the properties belong to
they can always find out the interfaces security. And are the best way to make
sure security is being done correctly.

Change-Id: I9572a1f6b1c219ce1ce2fc93739dcf344b162ffc